updated buildkite detectors
